+++ Quests/Kanto/P13RocketCeladonQuest.lua
@@ -86,7 +86,7 @@ function P13RocketCeladonQuest:isDone()
 end
 
 function P13RocketCeladonQuest:CeladonCity()
-	if self:needPokecenter() or not game.isTeamFullyHealed() or self.registeredPokecenter ~= "Pokecenter Celadon" or self.needCutPokemonFromBoxes then
+	if self:needPokecenter() or self:needsTeamHealingForStory() or self.registeredPokecenter ~= "Pokecenter Celadon" or self.needCutPokemonFromBoxes then
 		sys.debug("quest", "Going to heal Pokemon.")
 		return moveToCell(52, 19)